HS-753: Asynchronously update tasks and config changes (#574)

+++ inventory/src/test/java/org/opennms/horizon/inventory/service/taskset/response/DetectorResponseServiceIntTest.java
@@ -116,7 +116,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
         MonitoredServiceType relatedType = monitoredService.getMonitoredServiceType();
         assertEquals(monitoredServiceType, relatedType);
 
-        assertEquals(2, testGrpcService.getTimesCalled());
+        assertEquals(2, testGrpcService.getTimesCalled().intValue());
     }
 
     @Test
@@ -153,7 +153,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
         MonitoredServiceType relatedType = monitoredService.getMonitoredServiceType();
         assertEquals(monitoredServiceType, relatedType);
 
-        assertEquals(numberOfCalls*2, testGrpcService.getTimesCalled());
+        assertEquals(numberOfCalls*2, testGrpcService.getTimesCalled().intValue());
     }
 
     @Test
@@ -173,7 +173,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
         List<MonitoredService> monitoredServices = monitoredServiceRepository.findAll();
         assertEquals(0, monitoredServices.size());
 
-        assertEquals(0, testGrpcService.getTimesCalled());
+        assertEquals(0, testGrpcService.getTimesCalled().intValue());
     }
 
     @Test
@@ -220,7 +220,7 @@ class DetectorResponseServiceIntTest extends GrpcTestBase {
             assertEquals(TEST_TENANT_ID, monitoredService.getTenantId());
         }
 
-        assertEquals(numberOfCalls*2, testGrpcService.getTimesCalled());
+        assertEquals(numberOfCalls*2, testGrpcService.getTimesCalled().intValue());
 
         List<PublishTaskSetRequest> grpcRequests = testGrpcService.getRequests();
         assertEquals(monitorTypes.length*2, grpcRequests.size());
